Plans now can have decimal prices and intervals of every type like day, week, month and year. Plans and Product after creation on stripe updating synchronously, not using webhooks (#45). Webhook crashing issue on signature verification resolved (#46). These links used as a template base, not for real pages (#47). Case handled when the desired record not found (#48). Subscription will not be created when payment fails (#49). Context (#27).

+++ app/services/stripe_webhook_manager.rb
@@ -21,26 +21,22 @@ class StripeWebhookManager
       Stripe::Webhook.construct_event(
         payload, sig_header, ENV['STRIPE_ENDPOINT_SECRET']
       )
-    rescue JSON::ParserError, Stripe::SignatureVerificationError
+    rescue JSON::ParserError, Stripe::SignatureVerificationError => e
       # Invalid payload, Invalid signature
-      render body: nil, status: :bad_request
-      return
+      { body: nil, status: :bad_request, message: e.message }
     end
   end
 
   # Different Events Handling
   def handle_events
     case @params[:type]
-    when 'product.created'
-      update_app_product
-    when 'plan.created'
-      update_app_plan
     when 'payment_intent.payment_failed'
-      run_the_job
+      cancel_payment_intent
+      change_user_status
     when 'payment_intent.succeeded'
       update_user
     when 'customer.subscription.updated'
-      update_subscription
+      set_up_subscription
     when 'customer.subscription.deleted'
       delete_app_subscription
     when 'customer.updated'
@@ -52,31 +48,41 @@ class StripeWebhookManager
     end
   end
 
-  def update_app_product
-    app_product_name = @params[:data][:object][:name]
-    product = Product.find_by_name(app_product_name)
-    product.update!(stripe_id: @params[:data][:object][:id])
+  def cancel_payment_intent
+    intent_id = @params[:data][:object][:id]
+    Stripe::PaymentIntent.cancel(intent_id)
+  rescue Stripe::InvalidRequestError => e
+    { message: e.message }
   end
 
-  def update_app_plan
-    app_plan_name = @params[:data][:object][:nickname]
-    plan = Plan.find_by_nickname(app_plan_name)
-    plan.update!(stripe_id: @params[:data][:object][:id])
-  end
+  def change_user_status
+    user_email =
+      @params[:data][:object][:last_payment_error][:payment_method][:billing_details][:email]
+    return unless user_email.present?
 
-  def run_the_job
-    intent_id = params[:data][:object][:id]
-    PaymentIntentJob.perform_later(intent_id)
-    redirect_to(root_path)
+    user = User.find_by_email(user_email)
+    return unless user.present?
+
+    user.update(payment_status: false)
   end
 
   def update_user
     stripe_customer =
       @params[:data][:object][:charges][:data][0][:customer]
+    return unless stripe_customer.present?
+
     customer_object = Stripe::Customer.retrieve(stripe_customer)
     user_email = customer_object.email
     user = User.find_by_email(user_email)
     user.update!(payment_status: true) if user.present?
+  end
+
+  def set_up_subscription
+    if fetch_subscription.present?
+      update_subscription
+    else
+      create_subscription
+    end
   end
 
   def update_subscription
@@ -95,13 +101,39 @@ class StripeWebhookManager
     end
   end
 
+  def create_subscription
+    user_id = @params[:data][:object][:metadata][:user]
+    user = User.find(user_id)
+    return unless user.present?
+
+    return unless @params[:data][:object][:status] == 'active'
+
+    subscription = user.build_subscription(subscription_params)
+    subscription.update!(
+      stripe_id: @params[:data][:object][:id],
+      cancel_at_period_end: @params[:data][:object][:cancel_at_period_end],
+      current_period_start: @params[:data][:object][:current_period_start],
+      current_period_end: @params[:data][:object][:current_period_end]
+    )
+  end
+
+  def subscription_params
+    {
+      plan_id: @params[:data][:object][:metadata][:plan_id]
+    }
+  end
+
   def delete_app_subscription
+    return unless fetch_subscription.present?
+
     fetch_subscription.user.update!(payment_status: false)
     fetch_subscription.destroy!
   end
 
   def update_customer
     customer = User.find_by_stripe_customer_id(@params[:data][:object][:id])
+    return unless customer.present?
+
     customer.update!(
       stripe_payment_method_id:
       @params[:data][:object][:invoice_settings][:default_payment_method]
@@ -115,7 +147,7 @@ class StripeWebhookManager
   def update_user_payment_method
     user_id = @params[:data][:object][:client_reference_id]
     user = User.find_by_id(user_id)
-    return unless user
+    return unless user.present?
 
     setup_intent =
       Stripe::SetupIntent.retrieve(@params[:data][:object][:setup_intent])
